backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiofiles==23.2.1
openai==1.13.3
faiss-cpu==1.7.4
PyMuPDF==1.23.22
tiktoken==0.6.0
pydantic-settings==2.2.1
email-validator==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
websockets==12.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.28
langchain-text-splitters==0.0.1
numpy==1.26.4
redis==5.0.2
prometheus-client==0.20.0
structlog==24.1.0
alembic==1.13.1
